cqfd: add new ini parser
The .ini parser is implemented using a series of bash substitutions that
transform an .ini-style configuration file into shell instructions. The
original code was written by Andrés J. Díaz and later improved and
simplified.

In practice, sections are converted into functions whose names are
prefixed with cfg_section_, while properties are translated into
variables. The resulting code is then evaluated with eval, and the
section functions are executed to set the corresponding properties in
the shell environment.

That bash-substitutions based parser assumes lines are shell variables,
and thus, they are **NOT** commands. This assumption is incorrect and
dangerous.

The following .cqfdrc will destroy your world, **DO NOT TRY IT!**:

	[project]
	org="savoirfairelinux"
	name="cqfd"
	echo "Good Bye, World!"

	[build]
	command="/bin/echo" rm -Rf / 2>&1
	files=TO-LET-FUNCTION-RETURN-0

This configuration produces the following shell code:

	function cfg_section_project {
		org="savoirfairelinux"
		name="cqfd"
		echo "Good Bye, World!"
	}

	function cfg_section_build {
		command="/bin/echo" rm -Rf / 2>&1
		files=TO-LET-FUNCTION-RETURN-0
	}

When cqfd calls cfg_section_project(), it prints “Good Bye, World!” to
stdout. When it calls cfg_section_build(), it starts removing files from
the root directory until interrupted.

	$ cqfd
	Good Bye, World!
	rm: cannot remove '/usr/bin/[': Permission denied
	rm: cannot remove '/usr/bin/addpart': Permission denied
	rm: cannot remove '/usr/bin/apt': Permission denied
	rm: cannot remove '/usr/bin/apt-cache': Permission denied
	rm: cannot remove '/usr/bin/apt-cdrom': Permission denied
	rm: cannot remove '/usr/bin/apt-config': Permission denied
	rm: cannot remove '/usr/bin/apt-get': Permission denied
	rm: cannot remove '/usr/bin/apt-key': Permission denied
	rm: cannot remove '/usr/bin/apt-mark': Permission denied
	^C

While the echo command is easy to spot, the second command is far more
subtle and dangerous.

This issue is well known [1][2][3][4] and cannot be safely fixed in the
current approach.

This change introduces a completely new .ini parser that treats
configuration files as data and prevents command evaluation.

[1]: https://ajdiaz.wordpress.com/2008/02/09/bash-ini-parser/#comment-454
[2]: https://ajdiaz.wordpress.com/2008/02/09/bash-ini-parser/#comment-455
[3]: https://ajdiaz.wordpress.com/2008/02/09/bash-ini-parser/#comment-457
[4]: https://ajdiaz.wordpress.com/2008/02/09/bash-ini-parser/#comment-458