cqfd: run sudo if not in docker group
The docker client requires to run as root in rootfull mode. The docker
group allows a user to access to the Unix socket created by the docker
daemon.

The user needs to run docker client with more privileges (i.e. sudo) if
it is not part of that group[1].

This prepends sudo to CQFD_DOCKER automatically if CQFD_RUN_WITH_SUDO is
set to true. It turns the internal variable cqfd_docket into an array to
split the command into pieces ("sudo docker" -> "sudo" and "docker").

Note: ls and gc commands run docker, thus, both commands require to call
load_config() that sets the cqfd_docker "binary" to exec if user is not
in the docker group and if CQFD_RUN_WITH_SUDO is set.

[1]: https://docs.docker.com/install/linux/linux-postinstall/#manage-docker-as-a-non-root-user