Support for compatibility level
As every piece of software, cqfd is subject to breaking changes, either
because of features are fixed, or because of features are deprecated.

And, cqfd may need to maintain compatibility with the previous versions.

The .cqfdrc file is ususally committed in the projects using it; that
file is closely related to a specific version of cqfd.

To overcome the breaking changes, multiple versions of cqfd can be
installed in the same system; every cqfd verion is suffixed by its
compatibility level. For example:
 - cqfd-56 (5.6.x)
 - cqfd-57 (5.7.x)
 - cqfd-60 (6)
 - cqfd-70 (7)
 - cqfd-2026 (2026)?
 - ... as long as the compatibility level increases

cqfd is a symlink to the latest (or a specific) version.

Also, cqfd can maintain breaking changes in its code by checking the
compatibility level at runtime, either using the environment CQFD_COMPAT
(ex. export CQFD_COMPAT=56) or using the final filename (ex. cqfd ->
cqfd-56 (5.6.x)).

As an example, cqfd has fixed extra arguments containing whitespaces
since the commit e8bc9afb3381800c7f92045c6aa26f167802baa6 (5.5.0). This
commit allows to "revert" that breaking change:

cqfd succeeds with compatibility level >= 55

	$ env CQFD_COMPAT=55 CQFD_EXTRA_RUN_ARGS='--env MESSAGE="Hello,\ World!"' cqfd .shell env | grep MESSAGE
	MESSAGE="Hello, World!"

cqfd fails with compatibility level < 55

	$ env CQFD_COMPAT=54 CQFD_EXTRA_RUN_ARGS='--env MESSAGE="Hello,\ World!"' cqfd shell env | grep MESSAGE
	docker: invalid reference format: repository name (library/World!") must be lowercase

	Run 'docker run --help' for more information

This adds support for compatibility level, the cqfd's compatibility
level is set by the environment (CQFD_COMPAT) or deduced from script
filename.