cqfd: rework output of legacy variable message
The output of the fatal message is as follow is using a legacy variable:

	Warning: CQFD_EXTRA_VOLUMES is no more supported, use
			CQFD_EXTRA_RUN_ARGS="-v <local_dir>:<container_dir>"

The second line is prefixed by a series of blanks that depends on the
code indentation.

This reworks the output of the fatal message if using a legacy variable
so it appears in a single line by closing the quote and adding an espace
character.

	Warning: CQFD_EXTRA_VOLUMES is no more supported, use CQFD_EXTRA_RUN_ARGS="-v <local_dir>:<container_dir>"