tests: create temporary directory to /tmp/cqfd-test.XXXXXX
The temporary test directory is created to /tmp/tmp.XXXXXX and it is
removed on exit.

Some programs create files or directories with the exact same template.

This explicits the temporary test directory belongs to cqfd by updating
the file template to /tmp/cqfd-test.XXXXXX. It is easier to find the
remaining temporary directories not removed this way.