tests: 03-cqfd_{help,version}: create tmp files in TDIR
mktemp creates a files to /tmp/tmp.XXXXXX if TEMPLATE is not given and
if TMPDIR is unset.

The tests are run in a mktemp directory already[1][2], and that
directory is removed at the end of tests[3] (if the tests are not
interrupted[3]). The directory is pointed by "$TDIR".

Thus, additionnal temporary files and directories can be created in that
directory to avoid having multiple places to clean up during a test or
after the tests. It polutes even more the temporary directory of the
system if files and directories are left uncleaned.

This removes the use of mktemp to create files directly in the existing
temporary directory "$TDIR" to avoid poluting the temporary directory of
the system if the tests are interrupted.

[1]: https://github.com/savoirfairelinux/cqfd/blob/v5.7.0/tests/Makefile#L14
[2]: https://github.com/savoirfairelinux/cqfd/blob/v5.7.0/tests/Makefile#L17
[3]: https://github.com/savoirfairelinux/cqfd/blob/v5.7.0/tests/Makefile#L29-L31