tests: 05-cqfd_init_custom_image: rewrite pretest code
The code testing if the custom image does not exist before running the
test creating it **IS NOT** an actual test.

It is a safeguard to ensure the custom image is not already present and
to make sure the test actually creates it.

Note: The custom image remains created if the tests are interrupted
before the created image is cleaned up. Nevertheless, the risks are
mitigated by the use of the double $RANDOM suffix in the custom image
name, which "guarantee" to run multiple batches of test at the same
time.

This rewrites the code and transforms it as a none test case code, as in
05-cqfd_run_c_flavor[1].

[1]: https://github.com/savoirfairelinux/cqfd/blob/master/tests/05-cqfd_run_c_flavor#L15-L20