tests: 05-cqfd_init_custom_image: test pulling official image
cqfd with custom_img_name pulls the image if it does not exist in the
local registry; bypassing cqfd init.

Note: cqfd init builds the custom_img_name using the Dockerfile in the
.cqfd directory.

This adds a test to test the pulling of a official image (Ubuntu 16.04)
as the custom image set in custom_img_name image (the official image is
previously removed at setup).