tests: 03-cqfd_error: remove duplicated test
The test was introduced since commit[1]. It was later "superseded" by
the commit[2][3] and looks like that nowadays[4].

This removes the duplicated former test and keeps the later that is
tested very early in the test plan (first test).

Note: It avoids poluting even more the temporary directory of the system
if the tests are interrupted.

[1]: https://github.com/savoirfairelinux/cqfd/commit/d2954e71f66707f51f1a97b59b2ddf0023cb919a
[2]: https://github.com/savoirfairelinux/cqfd/commit/b65f8b0706bc42e94597be0adb756fca52cc84a5
[3]: https://github.com/savoirfairelinux/cqfd/blob/b65f8b0706bc42e94597be0adb756fca52cc84a5/tests/00-setup#L11-L14
[4]: https://github.com/savoirfairelinux/cqfd/blob/1f5f3be2f525ba06409f57658358e5f9d9228458/tests/00-setup#L13-L21
[5]: https://github.com/savoirfairelinux/cqfd/commit/da9d7e0c165ab5ab3a81058b3ee9abd8cc2d1063