cqfd: docker_run: remove implementation details in banner
The implementation details has not changed since 10 years and it is now
outdated. This is the magic of cqfd and it should be part of the
README.md or be part of cqfd(1).

This removes the outdated documentation of the function docker_run() in
its banner.

See commit 5d727f0273fb67a7dc6d0b2a287c6062ddeeb860.